% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nymphaea.R
\docType{data}
\name{nymphaea}
\alias{nymphaea}
\title{Example input file of Nymphaea ISSR data, taken from Reid et. al. (2021). This dataset has already been consolidated, and can be used
as input for the generation of an nMDS plot.}
\format{
A dataframe with columns for loci, and rows of replicate pairs. Grouping information is in the second column.
}
\usage{
data(nymphaea)
}
\description{
Example input file of Nymphaea ISSR data, taken from Reid et. al. (2021). This dataset has already been consolidated, and can be used
as input for the generation of an nMDS plot.
}
\examples{
data(nymphaea)
mat = nymphaea
group.names(mat)
}
\keyword{datasets}
